Added support for many-to-many associations where the foreign key of the join table references an alternative key. Also changed the implementation of "decimal" from Float to Double.

+++ modules/BusinessDomainDsl/org.mod4j.dsl.businessdomain.generator/src/main/java/org/mod4j/businessdomain/generator/helpers/BusinessClassHelpers.java
@@ -201,7 +201,7 @@ public class BusinessClassHelpers {
     }
 
     public static String javaNullableType(DecimalProperty p) {
-        return "Float";
+        return "Double";
     }
 
     public static String javaDefaultValue(DecimalProperty p) {
